requirements.txt
streamlit
Flask
feedparser
beautifulsoup4
requests
transformers
torch
nltk
gtts
langdetect
pydub
deep-translator
tavily-python
python-dotenv
fastapi
uvicorn
pydantic
spacy
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
